Move util functions to own file
